+++ StarmapBetterZoom/ZoomFix.cs
@@ -1,5 +1,5 @@
 ﻿using HarmonyLib;
-using PulsarPluginLoader.Patches;
+using PulsarModLoader.Patches;
 using System.Collections.Generic;
 using System.Reflection.Emit;
